requirements.txt
django-sekizai
coverage
unittest-xml-reporting==1.2
Sphinx
